Atualização Geral

+++ application/models/Pedido_model.php
@@ -1,6 +1,4 @@
-<?php
-
-defined('BASEPATH') OR exit('No direct script access allowed');
+<?php defined('BASEPATH') OR exit('No direct script access allowed');
 
 class Pedido_model extends CI_Model {
 
@@ -14,11 +12,6 @@ class Pedido_model extends CI_Model {
     public $columns;
     public $recordsTotal;
     public $recordsFiltered;
-
-    public function __construct()
-    {
-        parent::__construct();
-    }
 
     /**
      * Método responsável por solicitar um pedido
@@ -120,36 +113,36 @@ class Pedido_model extends CI_Model {
                         if ($vl->id_grupo_fk == "3") {
                             $benef['vl_total'][$i] = ($vl->vl_unitario*$vl->qtd_diaria);
                         } else {
-                            $benef['vl_total'][$i] = ($vl->vl_unitario*$vl->qtd_dia*$vl->qtd_diaria);
+                            $benef['vl_total'][$i] = ($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria));
                         }
-                        $benef['vl_repasse'][$i]      = isset($vl->vl_repasse) && $vl->vl_repasse != "" ? (($vl->vl_repasse*($vl->vl_unitario*$vl->qtd_dia*$vl->qtd_diaria))/100) : 0;
+                        $benef['vl_repasse'][$i]      = isset($vl->vl_repasse) && $vl->vl_repasse != "" ? (($vl->vl_repasse*($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria)))/100) : 0;
                         $benef['vl_rep_func'][$i]     = isset($vl->vl_rep_func) && $vl->vl_rep_func != "" ? $vl->vl_rep_func : 0;
 
                         # Devidir Grupos para calcular taxas
                         # Grupo Vale Transporte
                         if ($vl->id_grupo_fk == "1") {
-                            $benef['vl_total_vt'][$i] = ($vl->vl_unitario*$vl->qtd_dia*$vl->qtd_diaria);
+                            $benef['vl_total_vt'][$i] = ($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria));
                         } else {
                             $benef['vl_total_vt'][$i] = 0;
                         }
 
                         # Grupo Vale Refeição
                         if ($vl->id_grupo_fk == "2") {
-                            $benef['vl_total_cr'][$i] = ($vl->vl_unitario*$vl->qtd_dia*$vl->qtd_diaria);
+                            $benef['vl_total_cr'][$i] = ($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria));
                         } else {
                             $benef['vl_total_cr'][$i] = 0;
                         }
 
                         # Grupo Vale Alimentação
                         if ($vl->id_grupo_fk == "3") {
-                            $benef['vl_total_ca'][$i] = ($vl->vl_unitario*$vl->qtd_diaria);
+                            $benef['vl_total_ca'][$i] = ($vl->vl_unitario*($vl->qtd_diaria));
                         } else {
                             $benef['vl_total_ca'][$i] = 0;
                         }
 
                         # Grupo Vale Combustivel
                         if ($vl->id_grupo_fk == "4") {
-                            $benef['vl_total_cc'][$i] = ($vl->vl_unitario*$vl->qtd_dia*$vl->qtd_diaria);
+                            $benef['vl_total_cc'][$i] = ($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria));
                         } else {
                             $benef['vl_total_cc'][$i] = 0;
                         }
@@ -165,7 +158,7 @@ class Pedido_model extends CI_Model {
                         }
                         $this->db->insert('tb_item_pedido', $item);
                         $id_last_item_pedido = $this->db->insert_id();
-                        
+
                         # Salvar em tb_relatorio
                         $rel['id_pedido_fk']         = $valores->id;
                         $rel['id_empresa_fk']        = $valores->id_cliente;
@@ -245,6 +238,75 @@ class Pedido_model extends CI_Model {
     }
 
     /**
+     * Método responsável por exportar a visualizacao de um pedido
+     *
+     * @method exportVerPedido
+     * @param obj $valores Dados do pedido
+     * @access public
+     * @return obj Itens do Pedido
+     */
+    public function exportVerPedido($valores)
+    {
+        # Atribuir vars
+        $retorno = new stdClass();
+        $listas  = array();
+
+        # Buscar Itens Beneficios
+        if (!empty($valores->id_func) && is_array($valores->id_func)):
+            foreach ($valores->id_func as $value):
+                $this->db->select('b.id_beneficio_pk, b.vl_unitario, b.qtd_diaria, b.id_grupo_fk, b.id_item_beneficio_fk,
+                                   i.descricao, i.vl_rep_func, i.vl_repasse, pe.qtd_dia, vwf.cnpj, vwf.nome_razao, vwf.cpf,
+                                   vwf.rg, vwf.nome, c.num_cartao');
+                $this->db->from('tb_beneficio b');
+                $this->db->join('tb_item_beneficio i', 'b.id_item_beneficio_fk = i.id_item_beneficio_pk', 'inner');
+                $this->db->join('vw_funcionario vwf', 'vwf.id_funcionario_pk = b.id_funcionario_fk', 'inner');
+                $this->db->join('tb_periodo pe', 'pe.id_periodo_pk = vwf.id_periodo_pk', 'inner');
+                $this->db->join('tb_cartao c', 'c.id_funcionario_fk = b.id_funcionario_fk AND c.id_beneficio_fk = b.id_beneficio_pk', 'left');
+                $this->db->where('b.id_funcionario_fk', $value);
+                $resp = $this->db->get()->result();
+
+                if (!empty($resp)):
+                    $vl_total     = 0;
+                    $qtd_unitaria = 0;
+                    foreach ($resp as $vl):
+                        if ($vl->id_grupo_fk == "3") {
+                            $vl_total     = ($vl->vl_unitario*$vl->qtd_diaria);
+                            $qtd_unitaria = $vl->qtd_diaria;
+                        } else {
+                            $vl_total     = ($vl->vl_unitario*($vl->qtd_dia*$vl->qtd_diaria));
+                            $qtd_unitaria = ($vl->qtd_dia*$vl->qtd_diaria);
+                        }
+
+                        $lista                = new stdClass();
+                        $lista->cnpj          = $vl->cnpj;
+                        $lista->nome_razao    = $vl->nome_razao;
+                        $lista->cpf           = $vl->cpf;
+                        $lista->rg            = $vl->rg;
+                        $lista->nome          = $vl->nome;
+                        $lista->cod_benefico  = $vl->id_item_beneficio_fk;
+                        $lista->desc_benefico = $vl->descricao;
+                        $lista->vl_unitario   = "R$ ".number_format($vl->vl_unitario, 2, ',', '.');
+                        $lista->qtd_unitaria  = $qtd_unitaria;
+                        $lista->vl_total      = "R$ ".number_format($vl_total, 2, ',', '.');
+                        $lista->num_cartao    = $vl->num_cartao;
+                        $listas[]             = $lista;
+                    endforeach;
+                endif;
+            endforeach;
+            $retorno->status = TRUE;
+            $retorno->msg    = "Dados Gerados com Sucesso!";
+            $retorno->dados  = $listas;
+        else:
+            $retorno->status = FALSE;
+            $retorno->msg    = "Houve um erro ao Finalizar! Obrigat&oacute;rio cadastrar primeiro o(s) Benef&iacute;cio(s) do(s) Funcion&aacute;rio(s).";
+            $retorno->dados  = NULL;
+        endif;
+
+        # retornar
+        return $retorno;
+    }
+
+    /**
      * Método responsável por pesquisar e buscar pedidos
      *
      * @method getPedidoAcompanha
@@ -296,11 +358,10 @@ class Pedido_model extends CI_Model {
         # Consultar pedidos
         $this->db->select("p.id_pedido_pk, p.id_empresa_fk, e.cnpj, e.nome_razao, p.id_end_empresa_fk,
                            p.dt_pgto, CONCAT(DATE_FORMAT(p.dt_ini_beneficio, '%d/%m/%Y'), ' a ', DATE_FORMAT(p.dt_fin_beneficio, '%d/%m/%Y')) AS periodo, p.vl_itens,
-                           p.vl_taxa, p.vl_repasse, p.vl_total, p.id_status_pedido_fk, s.status_pedido, p.boleto_gerado, p.dt_hr_solicitacao, b.nome_boleto", FALSE);
+                           p.vl_taxa, p.vl_repasse, p.vl_total, p.id_status_pedido_fk, s.status_pedido, p.boleto_gerado, p.dt_hr_solicitacao", FALSE);
         $this->db->from('tb_pedido p');
         $this->db->join('tb_empresa e', 'p.id_empresa_fk = e.id_empresa_pk', 'inner');
         $this->db->join('tb_status_pedido s', 'p.id_status_pedido_fk = s.id_status_pedido_pk', 'inner');
-        $this->db->join('tb_boleto b', 'p.id_pedido_pk = b.id_pedido_fk', 'left');
         $this->db->where('p.id_empresa_fk', $this->session->userdata('id_client'));
         if (!empty($filter)):
             $where = implode(" OR ", $filter);
@@ -316,12 +377,30 @@ class Pedido_model extends CI_Model {
         if (!empty($resp_dados)):
 
             foreach ($resp_dados as $value):
+                # Buscar Nome Boleto
+                $this->db->select("nome_boleto");
+                $this->db->from('tb_boleto');
+                $this->db->where('id_pedido_fk', $value->id_pedido_pk);
+                $row_boleto = $this->db->get()->result();
+
+                $nome_boleto = NULL;
+                if (!empty($row_boleto)):
+                    $nome_boleto = $row_boleto[0]->nome_boleto;
+                endif;
+
                 # Botao
-                $id_pedido  = $value->id_pedido_pk;
-                $nome_boleto = $value->nome_boleto;
+                $id_pedido       = $value->id_pedido_pk;
+                $url_boleto_pdf  = PATH_PROJ.'/assets/boletos/'.$nome_boleto;
+                $url_boleto_html = base_url('/pedido/remitirboletohtml/'.$id_pedido);
+                if ($nome_boleto != "" && file_exists($url_boleto_pdf)):
+                    $link_boleto = base_url('assets/boletos/'.$nome_boleto);
+                else:
+                    $link_boleto = $url_boleto_html;
+                endif;
+
                 # $url_boleto = base_url('./pedido/gerarboleto/'.base64_encode($id_pedido));
                 # $url_view   = base_url('./pedido/ver/'.$id_pedido);
-                $acao       = "<button type='button' class='btn btn-success btn-xs btn-acao' title='Remitir Boleto' onclick='Pedido.verBoleto(\"$id_pedido\")'><i class='glyphicon glyphicon-barcode' aria-hidden='true'></i></button>";
+                $acao       = "<button type='button' class='btn btn-success btn-xs btn-acao' title='Remitir Boleto' onclick='Pedido.verBoleto(\"$link_boleto\")'><i class='glyphicon glyphicon-barcode' aria-hidden='true'></i></button>";
                 $acao      .= "<button type='button' class='btn btn-primary btn-xs btn-acao' title='Visualizar Pedido' onclick='Pedido.exportPedido(\"$id_pedido\")'><i class='glyphicon glyphicon-eye-open' aria-hidden='true'></i></button>";
                 # $acao      .= "<button type='button' class='btn btn-danger btn-xs btn-acao' title='Excluir Per&iacute;odo' onclick='Pedido.del(\"$id_period\")'><i class='glyphicon glyphicon-remove' aria-hidden='true'></i></button>";
 
@@ -401,11 +480,10 @@ class Pedido_model extends CI_Model {
         # Consultar pedidos
         $this->db->select("p.id_pedido_pk, p.id_empresa_fk, e.cnpj, e.nome_razao, p.id_end_empresa_fk,
                            p.dt_pgto, CONCAT(DATE_FORMAT(p.dt_ini_beneficio, '%d/%m/%Y'), ' a ', DATE_FORMAT(p.dt_fin_beneficio, '%d/%m/%Y')) AS periodo, p.vl_itens,
-                           p.vl_repasse, p.vl_taxa, p.vl_total, p.id_status_pedido_fk, s.status_pedido, p.boleto_gerado, p.dt_hr_solicitacao, b.nome_boleto", FALSE);
+                           p.vl_repasse, p.vl_taxa, p.vl_total, p.id_status_pedido_fk, s.status_pedido, p.boleto_gerado, p.dt_hr_solicitacao", FALSE);
         $this->db->from('tb_pedido p');
         $this->db->join('tb_empresa e', 'p.id_empresa_fk = e.id_empresa_pk', 'inner');
         $this->db->join('tb_status_pedido s', 'p.id_status_pedido_fk = s.id_status_pedido_pk', 'inner');
-        $this->db->join('tb_boleto b', 'p.id_pedido_pk = b.id_pedido_fk', 'left');
         if (!empty($filter)):
             $where = implode(" OR ", $filter);
             $this->db->where($where);
@@ -420,13 +498,30 @@ class Pedido_model extends CI_Model {
         if (!empty($resp_dados)):
 
             foreach ($resp_dados as $value):
+                # Buscar Nome Boleto
+                $this->db->select("nome_boleto");
+                $this->db->from('tb_boleto');
+                $this->db->where('id_pedido_fk', $value->id_pedido_pk);
+                $row_boleto = $this->db->get()->result();
+
+                $nome_boleto = NULL;
+                if (!empty($row_boleto)):
+                    $nome_boleto = $row_boleto[0]->nome_boleto;
+                endif;
+
                 # Botao
-                $id_pedido   = $value->id_pedido_pk;
-                $id_status   = $value->id_status_pedido_fk;
-                # $nome_boleto = $value->nome_boleto;
-                # $url_boleto  = base_url('./pedido/gerarboleto/'.base64_encode($id_pedido));
+                $id_pedido       = $value->id_pedido_pk;
+                $id_status       = $value->id_status_pedido_fk;
+                $url_boleto_pdf  = PATH_PROJ.'/assets/boletos/'.$nome_boleto;
+                $url_boleto_html = base_url('/pedido/remitirboletohtml/'.$id_pedido);
+                if ($nome_boleto != "" && file_exists($url_boleto_pdf)):
+                    $link_boleto = base_url('assets/boletos/'.$nome_boleto);
+                else:
+                    $link_boleto = $url_boleto_html;
+                endif;
+
                 # $url_view    = base_url('./pedido/ver/'.$id_pedido);
-                $acao        = "<button type='button' class='btn btn-success btn-xs btn-acao' title='Remitir Boleto' onclick='Pedido.verBoleto(\"$id_pedido\")'><i class='glyphicon glyphicon-barcode' aria-hidden='true'></i></button>";
+                $acao        = "<button type='button' class='btn btn-success btn-xs btn-acao' title='Remitir Boleto' onclick='Pedido.verBoleto(\"$link_boleto\")'><i class='glyphicon glyphicon-barcode' aria-hidden='true'></i></button>";
                 $acao       .= "<button type='button' class='btn btn-warning btn-xs btn-acao' title='Editar Status do Pedido' onclick='Pedido.alterStatus(\"$id_pedido\", \"$id_status\")'><i class='glyphicon glyphicon-edit' aria-hidden='true'></i></button>";
                 $acao       .= "<button type='button' class='btn btn-primary btn-xs btn-acao' title='Visualizar Pedido' onclick='Pedido.exportPedido(\"$id_pedido\")'><i class='glyphicon glyphicon-eye-open' aria-hidden='true'></i></button>";
                 $acao       .= "<button type='button' class='btn btn-success btn-xs btn-acao' title='Valida&ccedil;&atilde;o de Cr&eacute;dito' onclick='Pedido.validaCredito(\"$id_pedido\")'><i class='glyphicon glyphicon-credit-card' aria-hidden='true'></i></button>";
@@ -531,18 +626,33 @@ class Pedido_model extends CI_Model {
     {
         # Atribuir vars
         $retorno = new stdClass();
+        $file    = NULL;
 
-        # SQL
-        $this->db->where('id_pedido_pk', $id);
-        $this->db->delete('tb_pedido');
+        # Verificar se boleto esta em Remessa
+        $this->db->select("r.arquivo");
+        $this->db->from("tb_boleto b");
+        $this->db->join("tb_remessa_mov m", "m.id_boleto_fk = b.id_boleto_pk", "inner");
+        $this->db->join("tb_remessa r", "r.id_remessa_pk = m.id_remessa_fk", "inner");
+        $this->db->where("b.id_pedido_fk", $id);
+        $rows = $this->db->get()->result();
 
-        if ($this->db->affected_rows() > 0) {
-            $retorno->status = TRUE;
-            $retorno->msg    = "Exclus&atilde;o realizada com Sucesso!";
-        } else {
+        if (!empty($rows)):
+            $file = $rows[0]->arquivo;
             $retorno->status = FALSE;
-            $retorno->msg    = "Houve um erro ao Excluir! Tente novamente...";
-        }
+            $retorno->msg    = "Houve um erro ao Excluir! Necess&aacute;rio excluir a Remessa <strong>{$file}</strong>";
+        else:
+            # SQL
+            $this->db->where('id_pedido_pk', $id);
+            $this->db->delete('tb_pedido');
+
+            if ($this->db->affected_rows() > 0) {
+                $retorno->status = TRUE;
+                $retorno->msg    = "Exclus&atilde;o realizada com Sucesso!";
+            } else {
+                $retorno->status = FALSE;
+                $retorno->msg    = "Houve um erro ao Excluir! Tente novamente...";
+            }
+        endif;
 
         # retornar
         return $retorno;
